requirements.txt
pyTelegramBotAPI==4.30.0
peewee==3.19.0
python-dotenv==1.2.1
schedule==1.2.2
python-dateutil==2.9.0.post0
pytz==2025.2